# openapi.yml
openapi: 3.0.3
info:
  title: Sudoku API
  version: v1
paths:
...
  /puzzles:
    x-controller: puzzles
    post:
      x-operation: createPuzzle
      tags:
        - Puzzles
      security:
        - cookieAuth: []
      summary: Create a Puzzle
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              required:
                - puzzledifficulty
                - originalnumbers
                - userenterednumbers
                - completed
              properties:
                puzzledifficulty:
                  type: integer
                originalnumbers:
                  type: array
                  items:
                    type: integer
                userenterednumbers:
                  type: array
                  items:
                    type: integer
                completed:
                  type: boolean
            example:
              puzzledifficulty: 1
              originalnumbers: [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]
              userenterednumbers: [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]
              completed: false
      responses: 
        201:
          description: Puzzle Created
        400:
          description: Invalid request.
        401:
          description: Not authenticated.
    patch:
      x-operation: updateIt
      tags:
        - Puzzles
      security:
        - cookieAuth: []
      summary:
        Updates a puzzle
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userenterednumbers:
                  type: array
                  items:
                    type: integer
                completed:
                  type: boolean 
      responses:
        200:
          description: puzzle updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Puzzle'
        400:
          description: Invalid request
        401:
          description: Not authenticated
        403:
          description: Access denied
        404:
          description: Not found
    delete:
      x-operation: deletePuzzle
      tags:
        - Puzzles
      security:
        - cookieAuth: []
      summary: Delete a puzzle
      responses:
        204:
          description: puzzle deleted
        401:
          description: not authenticate
        403:
          description: access denied
        404:
          description: puzzle not found
    get:
      x-operation: getPuzzle
      tags:
        - Puzzels
      security:
        - cookieAuth: []
      summary: Get a puzzle
      responses:
        200:
          description: got the puzzle
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Puzzle'
        401:
          description: not authenticate
        403:
          description: access denied
        404:
          description: puzzle not found
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sudokuSessionId
  schemas:
...
    Puzzle:
      type: object
      properties:
        puzzledifficulty:
          type: integer
        originalnumbers:
          type: array
          items:
            type: integer
        userenterednumbers:
          type: array
          items:
            type: integer
        completed:
          type: boolean 